session storage
website now doesn't require login function to be called every time we change a new page. continuing working on game.

have run into a bug where the player guesses and the writerec doesn't upload the player guess and only uploads the player turn. but the player turn is stuck at "p1". working to fix this bug.